my_workouts_bot/pyproject.toml
[tool.poetry]
name = "myworkoutsbot"
version = "0.1.0"
description = ""
authors = ["Miroslav Kuzin"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-telegram-bot = "^20.2"
pandas = "^2.0.1"
seaborn = "^0.12.2"
strenum = "^0.4.10"
sqlalchemy = "^2.0.12"
alembic = "^1.10.4"
aiofiles = "^23.1.0"
environs = "^9.5.0"
asyncpg = "^0.27.0"
pyngrok = "^6.0.0"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
pydantic = "^1.10.7"
ujson = "^5.7.0"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
sqlalchemy-stubs = "^0.4"
flake8 = "^6.0.0"
autoflake = "^2.1.1"
isort = "^5.12.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"